requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
PyJWT==2.8.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
Werkzeug==3.0.1
